requirements.txt
Flask==0.10.1
Flask-Mail==0.7.6
Flask-SQLAlchemy==1.0
Flask-WTF==0.9.5
Flask-WhooshAlchemy==0.55
Jinja2==2.7.3
MarkupSafe==0.23
SQLAlchemy==0.9.6
Tempita==0.5.2
WTForms==1.0.5
Werkzeug==0.9.6
Whoosh==2.6.0
blinker==1.3
chardet==2.2.1
decorator==3.4.0
flup==1.0.2
gunicorn==19.1.0
itsdangerous==0.24
lamson==1.3.4
pbr==0.9.0
pytz==2014.4
six==1.7.3
sqlalchemy-migrate==0.9.1
